backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis and Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OAuth
authlib==1.2.1
httpx==0.25.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
shortuuid==1.0.11
user-agents==2.2.0

# Monitoring (optional for production)
sentry-sdk[fastapi]==1.38.0

# HTTP client
requests==2.31.0
